requirements.txt
setuptools
click
flask
itsdangerous
jinja2
markupsafe
werkzeug
psycopg2-binary